ISO c90 with thick warnings

+++ include/cosmic/cosmic_io.h
@@ -2,6 +2,7 @@
 #define COSMIC_IO_H
 
 #include <stdlib.h>
+#include <stdio.h>
 
 #include <cosmic/cosmic_lib.h>
 
@@ -11,10 +12,10 @@ extern "C" {
 
 typedef struct cosmic_io cosmic_io_t;
 
-COSMIC_DLL int cosmic_io_read(cosmic_io_t *, char *, size_t);
-COSMIC_DLL int cosmic_io_write(cosmic_io_t *, const char *, size_t);
+COSMIC_DLL ssize_t cosmic_io_read(cosmic_io_t *, char *, size_t);
+COSMIC_DLL ssize_t cosmic_io_write(cosmic_io_t *, const char *, size_t);
 COSMIC_DLL int cosmic_io_close(cosmic_io_t *);
-COSMIC_DLL int cosmic_io_copy(cosmic_io_t *, cosmic_io_t *, size_t);
+COSMIC_DLL ssize_t cosmic_io_copy(cosmic_io_t *, cosmic_io_t *, size_t);
 
 #ifdef __cplusplus
 }
