+++ include/cosmic/cosmic_map.h
@@ -19,9 +19,11 @@ COSMIC_DLL void cosmic_map_free(cosmic_map_t *, cosmic_map_dealloc);
 COSMIC_DLL size_t cosmic_map_size(const cosmic_map_t *);
 COSMIC_DLL int cosmic_map_get_value(const cosmic_map_t *, const cosmic_any_t,
                               cosmic_any_t *);
-COSMIC_DLL int cosmic_map_add(cosmic_map_t *, cosmic_any_t, cosmic_any_t);
+COSMIC_DLL ssize_t cosmic_map_add(cosmic_map_t *, cosmic_any_t, cosmic_any_t);
 COSMIC_DLL int cosmic_map_remove(cosmic_map_t *, const cosmic_any_t, cosmic_pair_t *);
-// Type is cosmic_pair_t
+/**
+ *  Type is cosmic_pair_t
+ */
 COSMIC_DLL cosmic_iterator_t *cosmic_map_iterator(const cosmic_map_t *);
 
 #ifdef __cplusplus
